requirements.txt
paho-mqtt==1.6.1
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.8.0
eventlet==0.33.3
requests==2.31.0
geopy==2.4.0
folium==0.14.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.1
hbmqtt==0.9.6
